Add stack

+++ hash_table/hash_table.h
@@ -8,12 +8,12 @@ typedef struct {
 
 typedef struct {
     hash_table_row *rows; 
-    int capacity; 
-    int size;
+    size_t capacity; 
+    size_t size;
     size_t value_size;
 } hash_table; 
 
-hash_table init_hash_table(int capacity, size_t value_size);
+hash_table init_hash_table(size_t capacity, size_t value_size);
 void clean_hash_table(hash_table *ht);
 
 char hash_table_set(hash_table *ht, char *key, void *value);
